web/www-for-beginners/source/requirements.txt
Flask==2.3.2
PyJWT==2.7.0